Fix argument in un-invert
The argument of the `TestFunction.invert` method indicates if the default function output should be multiplied by -1. To un-invert the function, as you intent to do at the end of the `__call__` method of `PyScannerBit`, you thus need to set this argument to `False`. I can understand that this is unintuitive behaviour, it would make more sense to have the `invert` method be a toggle, instead of a setter-like function (as it currently is).

I have taken the liberty to make this change :-)